fix(lsp): remove deprecated struct and method for TextDocumentCodeActionRequest fix(lsp): delete CancelResponse struct and associated method refactor(lsp): integrate and consolidate TextDocumentCodeAction request and response methods perf(lsp): optimize request and response handling by consolidating into requests.go fix(lsp): remove redundant DidCloseTextDocument handling and methods cleanup(lsp): delete hover implementation and move Hover structs to requests.go refactor(lsp): combine DidSaveTextDocumentParamsNotification and DidCloseTextDocumentParamsNotification in notifications.go fix(lsp): ensure all response methods align with their request counterparts refactor(lsp): rename fields and methods for clarity and consistency style(lsp): remove unused imports and improve code formatting across multiple files
